// src/api/generated/models/targetRequestType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type TargetRequestType = typeof TargetRequestType[keyof typeof TargetRequestType];


export const TargetRequestType = {
  GRADE: 'GRADE',
  CLASS: 'CLASS',
  STUDENT: 'STUDENT',
} as const;